requirements_docker.txt
absl-py==2.4.0
accelerate==1.12.0
adagio==0.2.6
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
coreforecast==0.0.16
cycler==0.12.1
einops==0.8.2
einshape==1.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
fugue==0.9.4
gdown==5.2.1
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
h5py==3.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightning-utilities==0.15.2
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpi4py==4.1.1
mpmath==1.3.0
narwhals==2.16.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.0
packaging==26.0
pandas==3.0.0
patsy==1.0.2
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
sentry-sdk==2.52.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
statsforecast==2.0.1
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
timesfm @ git+https://github.com/google-research/timesfm.git@2dcc66fbfe2155adba1af66aa4d564a0ee52f61e
tokenizers==0.22.2
torchmetrics==1.8.2
tqdm==4.67.3
transformers==5.1.0
triad==1.0.1
typer==0.21.1
typer-slim==0.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.3
utilsforecast==0.2.15
wandb==0.24.2
